// src/common/enum/user.enum.ts
export enum GenderEnum {
  male = "male",
  female = "female",
}

export enum RoleEnum {
  user = "user",
  admin = "admin",
  superAdmin = "superAdmin",
}

export enum emailEnum {
  confirmEmail = "confirmEmail",
  resetPassword = "resetPassword",
}

export enum ProviderEnum {
  google = "google",
  local = "local",
}